requirements-llama.txt
# Core dependencies for Llama3 version
fastapi
uvicorn
pydantic
python-dotenv
twilio
together
